requirements.txt
blinker==1.8.2
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
python-dotenv==1.0.1
redis==5.0.6
requests==2.32.3
spotipy==2.24.0
urllib3==2.2.2
Werkzeug==3.0.3
